base: omnistyle: Stinky workaround for header shadow and tinted images
Issue: Some of our headers uses textures are tinted with android:tint, but also the headers are tinted with PorterDuff method to produce the shadow.
This situation makes the system accent tint went full black with even a 1% of 'shadow' and removes all the color from the headers (making them black, and looking fugly on dark themes)

So, as a bad alternative method, let's use a new view based on the qs background but fully black to work the same way as the previous shadow, but making the tint issue gone.

Signed-off-by: rayrzy <vincentrayananda@gmail.com>
Change-Id: I6b87a7e1f289549ef4e399f5ed528f6130153cfb